Be comment

+++ server/demo/src/main/java/example/domain/artistPost/service/artistPostService.java
@@ -1,9 +1,14 @@
 package example.domain.artistPost.service;
 
+import example.domain.artist.entity.Artist;
 import example.domain.artistPost.entity.ArtistPost;
 import example.domain.artistPost.repository.artistPostRepository;
+import example.domain.feedPost.entity.FeedPost;
 import example.global.exception.BusinessLogicException;
 import example.global.exception.ExceptionCode;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -13,35 +18,49 @@ import java.util.Optional;
 public class artistPostService {
     private artistPostRepository artistPostRepository;
 
-    public artistPostService(artistPostRepository artistPostRepository){
+    public artistPostService(artistPostRepository artistPostRepository) {
         this.artistPostRepository = artistPostRepository;
     }
 
-    public ArtistPost createArtistPost(ArtistPost artistPost){
+    public ArtistPost createArtistPost(ArtistPost artistPost) {
         return artistPostRepository.save(artistPost);
     }
 
-    public ArtistPost findArtistPost(int artistId){
+    public ArtistPost findArtistPost(int artistId) {
         return artistPostRepository.findById(artistId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.ARTISTPOST_NOT_FOUND));
     }
 
-    public ArtistPost updateArtistPost(ArtistPost artistPost){
-        ArtistPost findArtistPost = findArtistPost(artistPost.getId());
 
-        if(artistPost.getArtist().getArtistId() != findArtistPost.getArtist().getArtistId()) {
+    public ArtistPost updateArtistPost(ArtistPost artistPost) {
+        ArtistPost findArtistPost = findArtistPost(artistPost.getArtistPostId());
+
+        if (artistPost.getArtist().getArtistId() != findArtistPost.getArtist().getArtistId()) {
             throw new BusinessLogicException(ExceptionCode.ARTISTPOST_AUTHOR_NOT_MATCH);
-        }else{
+        } else {
             Optional.ofNullable(artistPost.getContent()).ifPresent(content -> findArtistPost.setContent(content));
             Optional.ofNullable(artistPost.getImg()).ifPresent(img -> findArtistPost.setImg(img));
-            findArtistPost.setModifiedAt(LocalDateTime.now());
+            findArtistPost.setCreatedAt(LocalDateTime.now());
         }
 
         return artistPostRepository.save(findArtistPost);
     }
 
-    public void deleteArtistPost(int artistId){
-        ArtistPost findArtistPost = findArtistPost(artistId);
+    public void deleteArtistPost(Artist artist, ArtistPost artistPost) {
+        ArtistPost findArtistPost = findArtistPost(artistPost.getArtistPostId());
+        if (artist.getArtistId() != findArtistPost.getArtist().getArtistId()) {
+            throw new BusinessLogicException(ExceptionCode.ARTISTPOST_AUTHOR_NOT_MATCH);
+        }
         artistPostRepository.delete(findArtistPost);
     }
 }
+/*
+    public Page<ArtistPost> findArtistPosts(int groupId, int page, int size){
+        Page<FeedPost> artistPosts = artistPostRepository.findAllByArtistGroupId(groupId, PageRequest.of(page, size, Sort.by("id").descending()));
+
+        return artistPosts;
+//        return artistPostRepository.findAll(PageRequest.of(page, size,Sort.by("id").descending()));
+    }
+}
+/*
+ */
